move test code to root
